Reorganize project while attempting to maintain as much backwards compatibility as possible
This commit is intended to reorganize the project to increase code reuse and make the majority of the behavior more consistent.  This implements changes so that every call, whether inside of a batch or async execution or using a database instance directly, will return a Job construct with a lazily loading result.  This enables the use of asynchronous http requests as well as other futures-based http responses (like multiprocessing.futures for example).  However, this is not enabled by default- it requires setting the old_behavior flag to False to maintain as much backwards compatibility as possible.

This commit does make a few changes.  Any behavioral changes (mostly around asynchronous job handling and user management) are marked with # TODO CHANGED .  The other potential changes involve differences in import statements.  I could use some guidance as to which classes are public facing and should be imported in the __init__.py file and which are intended to be internal.