Attempt 2 which addresses presented concerns in pr#64
This is a second attempt to reorganize this project so that it both can lazily load responses for interacting with the database using asynchronous requests and maintain the same api reference.

Something I've realized is that there are a lot of cases where the simplest way to organize the project for development diverges with the existing api, which is intended to be concise for use. So I've tried splitting up the two through creation of the arango_internals package. This passes all the original tests without changing imports at all. The changes which existed in the last pr to user management and async jobs still exist.

All of the changes suggested in the original pr, with the exception of the relative imports in __init__.py files which I believe are not an issue, have been implemented.

Thoughts?